Wait for a S3 bucket to be created before using it
This removes the change made in a1355917ecf75e1f1f37101505a351743d590121
that disables virtual host URIs by instead waiting for the bucket to be
created. Creating a bucket should be a reasonably infrequent operation,
and thus this won't happen particularly often.

Specifically waiting for a bucket creation to complete feels less
fragile than relying on the particular implementation details of S3, to
me.